spi: qspi: cannot use pointer type in shift operand directly
drivers/spi/spi-dw-quad.c: In function 'set_cs':
./include/vdso/bits.h:7:40: error: invalid operands to binary << (have 'long unsigned int' and 'u8 *' {aka 'unsigned char *'})
    7 | #define BIT(nr)                 (UL(1) << (nr))
      |                                        ^~
      |                                        |
      |                                        u8 * {aka unsigned char *}
drivers/spi/spi-dw-quad.c:228:44: note: in expansion of macro 'BIT'
  228 |                 dw_writel(dws, DW_SPI_SER, BIT(spi->chip_select));
      |                                            ^~~

Fixes: 5d2672305852 ("drivers/spi: Add THEAD TH1520 QSPI driver")

+++ drivers/spi/spi-dw-quad.c
@@ -225,7 +225,7 @@ static void __maybe_unused set_cs(struct spi_device *spi, bool enable)
 		chip->cs_control(!enable);
 
 	if (!enable)
-		dw_writel(dws, DW_SPI_SER, BIT(spi->chip_select));
+		dw_writel(dws, DW_SPI_SER, BIT((unsigned long)spi->chip_select));
 	else if (dws->cs_override)
 		dw_writel(dws, DW_SPI_SER, 0);
 }
